NFTCard - new component added
